@@ -60,6 +60,6 @@
     "react-dom": "^15.4.2"
   },
   "peerDependencies": {
-    "react": "0.14.x || 15.* || ^15.0.0"
+    "react": "^15.5.0"
   }
 }
